requirements.txt
runway-python==0.5.4
torch>=1.0
torchvision
tensorflow
tensorboard==1.14.0
terminaltables
pillow
